ollama_GUI/requirements.txt
PySide6>=6.6
requests
python-dotenv
psutil
pyautogui
Pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
